requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1
django-extensions==3.2.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
oauthlib==3.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
